API/requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
paho-mqtt
requests
